+++ packages/react/playwright/index.ts
@@ -1,4 +1,0 @@
-// Import styles, initialize component theme here.
-// import '../src/common.css';
-
-export {};