+++ packages/react/src/evm/connectors/magic.ts
@@ -1,3 +1,4 @@
+import { useConnect } from "../hooks/wagmi-required/useConnect";
 import { providers, utils } from "ethers";
 import type {
   LoginWithMagicLinkConfiguration,
@@ -5,7 +6,12 @@ import type {
   MagicSDKAdditionalConfiguration,
 } from "magic-sdk";
 import invariant from "tiny-invariant";
-import { Chain, Connector, normalizeChainId } from "wagmi";
+import {
+  Chain,
+  Connector,
+  normalizeChainId,
+  useContext as useWagmiContext,
+} from "wagmi";
 
 export interface MagicConnectorArguments
   extends MagicSDKAdditionalConfiguration {
@@ -185,4 +191,90 @@ export class MagicConnector extends Connector {
       window.localStorage.removeItem(LOCAL_STORAGE_KEY);
     }
   }
+}
+
+/**
+ * Hook for connecting to an email wallet using magic link.
+ * This enables users without their own wallets to connect to your application and sign transactions securely using their email.
+ *
+ * ```javascript
+ * import { useMagic } from "@thirdweb-dev/react"
+ * ```
+ *
+ *
+ * @example
+ * Before using this hook, you first need to set up the magic configuration in your `ThirdwebProvider`, including your magic API key.
+ *
+ * ```javascript
+ * // Add the magic configuration object to your wallet connectors
+ * const connectors = [
+ *   "metamask",
+ *   "walletConnect",
+ *   "walletLink",
+ *   {
+ *     name: "magic",
+ *     options: {
+ *       apiKey: "your-magic-api-key",
+ *     }
+ *   }
+ * ]
+ *
+ * // Add the above to the walletConnectors prop of your <ThirdwebProvider />
+ * const Provider = ({ children }) => (
+ *   return (
+ *     <ThirdwebProvider
+ *       walletConnectors={connectors}
+ *       // Specify remaining parameters
+ *       ...
+ *     >
+ *       {children}
+ *     </ThirdwebProvider>
+ *   )
+ * }
+ * ```
+ *
+ * In order to use the hook to connect users with magic link, you just need to provide the users email to the connect function.
+ *
+ * You can setup the hook with the following configuration:
+ * ```javascript
+ * import { useMagic } from "@thirdweb-dev/react"
+ * import { useState } from "react"
+ *
+ * const LoginWithMagicLink = () => {
+ *   const connectWithMagic = useMagic()
+ *   const [email, setEmail] = useState()
+ *
+ *   return (
+ *     <div>
+ *       <input value={email} onChange={(e) => setEmail(e.target.value)} />
+ *       <button onClick={() => connectWithMagic({ email })}>Login</button>
+ *     </div>
+ *   )
+ * }
+ * ```
+ *
+ * @public
+ */
+export function useMagic() {
+  const wagmiContext = useWagmiContext();
+  invariant(
+    wagmiContext,
+    `useMagic() can only be used inside <ThirdwebProvider />. If you are using <ThirdwebSDKProvider /> you will have to use your own wallet-connection logic.`,
+  );
+  const [connectors, connect] = useConnect();
+  if (connectors.loading) {
+    return () => Promise.reject("Magic connector not ready to be used, yet");
+  }
+
+  const connector = connectors.data.connectors.find((c) => c.id === "magic");
+
+  invariant(
+    connector,
+    "Magic connector not found, please make sure it is provided to your <ThirdwebProvider />",
+  );
+
+  return (configuration: LoginWithMagicLinkConfiguration) => {
+    (connector as MagicConnector).setConfiguration(configuration);
+    return connect(connector);
+  };
 }