+++ packages/react/src/evm/hooks/contracts/useEditionDrop.ts
@@ -1,4 +1,4 @@
-import { RequiredParam } from "../../../core/types/shared";
+import { RequiredParam } from "../../../core/query-utils/required-param";
 import { showDeprecationWarning } from "../../utils/deprecation-warning";
 import { useContract } from "../async/contracts";
 
@@ -27,8 +27,8 @@ import { useContract } from "../async/contracts";
  * @deprecated
  * This hook is deprecated and will be removed in a future major version. You should use {@link useContract} instead.
  * ```diff
- * - const editionDrop = await sdk.useEditionDrop("0x1234...");
- * + const editionDrop = await sdk.useContract("0x1234...", "edition-drop").contract;
+ * - const editionDrop = useEditionDrop("0x1234...");
+ * + const editionDrop = useContract("0x1234...", "edition-drop").contract;
  * ```
  */
 export function useEditionDrop(contractAddress: RequiredParam<string>) {