+++ packages/react/src/evm/providers/full.tsx
@@ -1,18 +1,12 @@
-import {
-  GnosisConnectorArguments,
-  GnosisSafeConnector,
-} from "../connectors/gnosis-safe";
-import { MagicConnector, MagicConnectorArguments } from "../connectors/magic";
+import type { GnosisSafeConnector } from "../connectors/gnosis-safe";
+import type { MagicConnector } from "../connectors/magic";
 import {
   Chain,
   SupportedChain,
   defaultSupportedChains,
 } from "../constants/chain";
 import { ThirdwebAuthConfig } from "../contexts/thirdweb-auth";
-import {
-  ThirdwebConfigProvider,
-  defaultChainRpc,
-} from "../contexts/thirdweb-config";
+import { ThirdwebConfigProvider } from "../contexts/thirdweb-config";
 import { ThirdwebSDKProvider, ThirdwebSDKProviderProps } from "./base";
 import { QueryClient } from "@tanstack/react-query";
 import {
@@ -20,6 +14,7 @@ import {
   getProviderForNetwork,
   SDKOptionsOutput,
 } from "@thirdweb-dev/sdk";
+import { DEFAULT_RPC_URLS } from "@thirdweb-dev/sdk";
 import type { ThirdwebStorage } from "@thirdweb-dev/storage";
 import React, { useMemo } from "react";
 import {
@@ -27,6 +22,7 @@ import {
   ProviderProps as WagmiproviderProps,
   useProvider,
   useSigner,
+  Connector,
 } from "wagmi";
 import { CoinbaseWalletConnector } from "wagmi/connectors/coinbaseWallet";
 import { InjectedConnector } from "wagmi/connectors/injected";
@@ -61,32 +57,12 @@ export type WalletLinkConnectorType =
 /**
  * @internal
  */
-export type MagicConnectorType =
-  | "magic"
-  | {
-      name: "magic";
-      options: Omit<MagicConnectorArguments, "network">;
-    };
-
-/**
- * @internal
- */
-export type GnosisConnectorType =
-  | "gnosis"
-  | {
-      name: "gnosis";
-      options: GnosisConnectorArguments;
-    };
-
-/**
- * @internal
- */
 export type WalletConnector =
   | InjectedConnectorType
   | WalletConnectConnectorType
   | WalletLinkConnectorType
-  | MagicConnectorType
-  | GnosisConnectorType;
+  | GnosisSafeConnector
+  | MagicConnector;
 
 /**
  * @internal
@@ -223,7 +199,7 @@ export const ThirdwebProvider = <
   TSupportedChain extends SupportedChain = SupportedChain,
 >({
   sdkOptions,
-  chainRpc = defaultChainRpc,
+  chainRpc = DEFAULT_RPC_URLS,
   supportedChains = defaultSupportedChains.map(
     (c) => c.id,
   ) as TSupportedChain[],
@@ -282,6 +258,9 @@ export const ThirdwebProvider = <
       connectors: ({ chainId }: { chainId?: number }) => {
         return walletConnectors
           .map((connector) => {
+            if (connector instanceof Connector) {
+              return connector;
+            }
             // injected connector
             if (
               (typeof connector === "string" &&
@@ -346,25 +325,7 @@ export const ThirdwebProvider = <
                       },
               });
             }
-            if (typeof connector === "object" && connector.name === "magic") {
-              const jsonRpcUrl = _rpcUrlMap[chainId || desiredChainId || 1];
-              return new MagicConnector({
-                chains: _supporrtedChains,
-                options: {
-                  ...connector.options,
-                  network: { rpcUrl: jsonRpcUrl, chainId: desiredChainId || 1 },
-                  rpcUrls: _rpcUrlMap,
-                },
-              });
-            }
-            if (
-              (typeof connector === "string" && connector === "gnosis") ||
-              (typeof connector === "object" && connector.name === "gnosis")
-            ) {
-              return new GnosisSafeConnector({
-                chains: _supporrtedChains,
-              });
-            }
+
             return null;
           })
           .filter((c) => c !== null);