+++ packages/storage/src/core/downloaders/storage-downloader.ts
@@ -20,25 +20,25 @@ export class StorageDownloader implements IStorageDownloader {
     gatewayUrls: GatewayUrls,
     attempts = 0,
   ): Promise<Response> {
-    // Replace recognized scheme with the highest priority gateway URL that hasn't already been attempted
-    let resolvedUri;
-    try {
-      resolvedUri = replaceSchemeWithGatewayUrl(uri, gatewayUrls, attempts);
-    } catch (err: any) {
-      // If every gateway URL we know about for the designated scheme has been tried (via recursion) and failed, throw an error
-      if (err.includes("[GATEWAY_URL_ERROR]")) {
-        throw new Error(
-          "[FAILED_TO_DOWNLOAD_ERROR] Unable to download from URI - all gateway URLs failed to respond.",
-        );
-      }
+    if (attempts > 3) {
+      throw new Error(
+        "[FAILED_TO_DOWNLOAD_ERROR] Failed to download from URI - too many attempts failed.",
+      );
+    }
 
-      throw err;
+    // Replace recognized scheme with the highest priority gateway URL that hasn't already been attempted
+    const resolvedUri = replaceSchemeWithGatewayUrl(uri, gatewayUrls, attempts);
+    // If every gateway URL we know about for the designated scheme has been tried (via recursion) and failed, throw an error
+    if (!resolvedUri) {
+      throw new Error(
+        "[FAILED_TO_DOWNLOAD_ERROR] Unable to download from URI - all gateway URLs failed to respond.",
+      );
     }
 
     const res = await fetch(resolvedUri);
 
     // If request to the current gateway fails, recursively try the next one we know about
-    if (!res.ok) {
+    if (res.status >= 500 || res.status === 403 || res.status === 408) {
       console.warn(
         `Request to ${resolvedUri} failed with status ${res.status} - ${res.statusText}`,
       );