+++ packages/auth/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @thirdweb-dev/auth
 
+## 2.0.23
+
+## 2.0.22
+
+## 2.0.21
+
+### Patch Changes
+
+- [#327](https://github.com/thirdweb-dev/js/pull/327) [`ef27aad`](https://github.com/thirdweb-dev/js/commit/ef27aad0aafc4577e85f44dc77dfbe880bd239b5) Thanks [@jnsdls](https://github.com/jnsdls)! - enable e2e testing
+
+## 2.0.20
+
+## 2.0.19
+
+## 2.0.18
+
 ## 2.0.17
 
 ## 2.0.16